+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Changed
 
+- removed error return in case of JobAlreadyExists in `create_job` function
 - update_job returns the updated job item
 - made create_job atomic to avoid race conditions
 - handle jobs in tokio tasks